+++ libregf/libregf_value.h
@@ -124,7 +124,7 @@ int libregf_value_get_utf16_name(
 LIBREGF_EXTERN \
 int libregf_value_get_value_type(
      libregf_value_t *value,
-     uint16_t *value_type,
+     uint32_t *value_type,
      libcerror_error_t **error );
 
 LIBREGF_EXTERN \
